Tidy up and comments
